+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Created By: BENGUERGOURA Oussama & HADJERSI Mohamed | 20/10/2024
 
 ### Added
 
+* Adding `BasicDialog` and `ConfirmationDialog` with his tests and documentation. [GTS-PRINT#7](https://quire.io/w/GTS-PRINT31/7)
 * Adding documentation for `BarSide`, `Input`, `DataTable`. [GTS-PRINT#14](https://quire.io/w/GTS-PRINT31/14)
 * Adding `ButtonsComponent` in `index.js` file. [GTS-PRINT#4](https://quire.io/w/GTS-PRINT31/4)
 * Adding `Buttons.spec.js` for testing `ButtonComponent` componement. [GTS-PRINT#4](https://quire.io/w/GTS-PRINT31/4)